krita/requirements.txt
imageio==2.16.1
joblib==1.1.0
llvmlite==0.38.0
networkx==2.7.1
numba==0.55.1
numpy==1.20.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyWavelets==1.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
threadpoolctl==3.1.0
tifffile==2022.3.16

parse
fire==0.2.1
ipython==6.5.0
matplotlib==3.0.3
pip==19.3.1
plac==0.9.6
py==1.6.0
torch
torchvision
tqdm==4.36.1
visdom==0.1.8.9
pathlib==1.0.1
pyyaml==5.3.1
pandas
seaborn
statsmodels
